requirements.txt
tqdm
pillow
sklearn
torch
torchvision
sklearn
albumentations

